Add files via upload
The "Video Metadata Editor and Player" is an app that helps users organize and find videos more easily by using metadata tags such as persons, events, and locations. With this app, users can save and edit video details in the database and within the video files, making it simple to keep the video gallery organized. The app allows users to search for specific video segments based on the metadata they have entered. This makes it easy to find particular scenes or moments in their video collection.